Addressed PR comments, cached mods parsed for nodes
